init engine

+++ kengine/src/generics/kcollide.rs
@@ -1,10 +1,10 @@
 use generics::kcoord::Kcoord;
 
-pub fn collide_with_window(a: &mut Kcoord, w: i64, h: i64, s: i64){
-    if a.x * s < 0 {a.x = w / s;}
-    if a.x * s > w {a.x = 0;}
-    if a.y * s < 0 {a.y = h / s;}
-    if a.y * s > h {a.y = 0;}
+pub fn collide_with_window(a: &mut Kcoord, w: f64, h: f64, s: f64){
+    if a.x * s < 0. {a.x = w / s;}
+    if a.x * s > w {a.x = 0.;}
+    if a.y * s < 0. {a.y = h / s;}
+    if a.y * s > h {a.y = 0.;}
 }
 
 pub fn collide_with_part(a: &Kcoord, b: &Vec<Kcoord>) -> i32{
