+++ kengine/src/generics/kcoord.rs
@@ -1,11 +1,11 @@
-#[derive(Clone)]
+#[derive(Debug, Clone, Copy)]
 pub struct Kcoord{
-    pub x: i64,
-    pub y: i64
+    pub x: f64,
+    pub y: f64
 }
 
 impl Kcoord{
-    pub fn new(x: i64, y: i64) -> Kcoord{
+    pub fn new(x: f64, y: f64) -> Kcoord{
         Kcoord{
             x: x,
             y: y
